+++ src/test/groovy/org/jenkinsci/gradle/plugins/jpi/JpiPomCustomizerSpec.groovy
@@ -2,6 +2,7 @@ package org.jenkinsci.gradle.plugins.jpi
 
 import org.custommonkey.xmlunit.XMLUnit
 import org.gradle.api.Project
+import org.gradle.api.internal.project.ProjectInternal
 import org.gradle.testfixtures.ProjectBuilder
 import spock.lang.Specification
 
@@ -21,6 +22,7 @@ class JpiPomCustomizerSpec extends Specification {
                 coreVersion = '1.580.1'
             }
         }
+        (project as ProjectInternal).evaluate()
 
         when:
         new JpiPomCustomizer(project).customizePom(pom)
@@ -53,6 +55,7 @@ class JpiPomCustomizerSpec extends Specification {
                 }
             }
         }
+        (project as ProjectInternal).evaluate()
 
         when:
         new JpiPomCustomizer(project).customizePom(pom)
@@ -70,6 +73,7 @@ class JpiPomCustomizerSpec extends Specification {
                 gitHubUrl = 'https://bitbucket.org/lorem/ipsum'
             }
         }
+        (project as ProjectInternal).evaluate()
 
         when:
         new JpiPomCustomizer(project).customizePom(pom)
@@ -89,6 +93,7 @@ class JpiPomCustomizerSpec extends Specification {
                 mavenLocal()
             }
         }
+        (project as ProjectInternal).evaluate()
 
         when:
         new JpiPomCustomizer(project).customizePom(pom)
@@ -108,6 +113,7 @@ class JpiPomCustomizerSpec extends Specification {
                 mavenCentral()
             }
         }
+        (project as ProjectInternal).evaluate()
 
         when:
         new JpiPomCustomizer(project).customizePom(pom)