+++ CHANGELOG.md
@@ -4,6 +4,8 @@
     configuration to recognize a non-default value
   * add a JAR dependency for each HPI/JPI dependency to the `testCompile` configuration
     ([JENKINS-17129](https://issues.jenkins-ci.org/browse/JENKINS-17129))
+  * added `configureRepositories` option to be able to skip configuration of repositories
+    ([JENKINS-17130](https://issues.jenkins-ci.org/browse/JENKINS-17130))
 
 ## 0.9.1 (2015-02-17)